Housekeeping
- Check user owns folder they are adding to or viewing
- Clippy improvements
- Remove `/test` endpoint
