+++ api/src/routes/flashcard.rs
@@ -4,32 +4,9 @@ use actix_web::{
     HttpRequest, HttpResponse, Responder,
 };
 use serde::{Deserialize, Serialize};
-use sqlx::PgPool;
 use ts_rs::TS;
 
-use crate::AppState;
-
-pub async fn get_folder_owner(folder_id: i32, db_pool: &PgPool) -> Option<i32> {
-    let top_level_folder = sqlx::query!(
-        "WITH RECURSIVE f AS(
-          SELECT id, parent_id FROM folder WHERE id = $1
-          UNION
-	        SELECT folder.id as id, folder.parent_id as parent_id FROM f, folder WHERE folder.id = f.parent_id
-        ) SELECT * FROM f WHERE parent_id IS NULL",
-        folder_id
-    ).fetch_one(db_pool).await.ok()?.id;
-
-    Some(
-        sqlx::query!(
-            "SELECT id FROM app_user WHERE flashcards = $1",
-            top_level_folder
-        )
-        .fetch_one(db_pool)
-        .await
-        .ok()?
-        .id,
-    )
-}
+use crate::{routes::folder::get_folder_owner, AppState};
 
 #[derive(Serialize, Deserialize, Debug, Clone, TS)]
 #[serde(rename_all = "camelCase")]
