+++ api/src/routes/folder.rs
@@ -1,21 +1,43 @@
 use actix_web::{
+    post,
     web::{self, Data},
-    HttpRequest, HttpResponse, Responder, post,
+    HttpRequest, HttpResponse, Responder,
 };
 use serde::{Deserialize, Serialize};
+use sqlx::PgPool;
 use ts_rs::TS;
 
 use crate::{utils, AppState};
+
+pub async fn get_folder_owner(folder_id: i32, db_pool: &PgPool) -> Option<i32> {
+    let top_level_folder = sqlx::query!(
+        "WITH RECURSIVE f AS(
+          SELECT id, parent_id FROM folder WHERE id = $1
+          UNION
+	        SELECT folder.id as id, folder.parent_id as parent_id FROM f, folder WHERE folder.id = f.parent_id
+        ) SELECT * FROM f WHERE parent_id IS NULL",
+        folder_id
+    ).fetch_one(db_pool).await.ok()?.id;
+
+    Some(
+        sqlx::query!(
+            "SELECT id FROM app_user WHERE flashcards = $1",
+            top_level_folder
+        )
+        .fetch_one(db_pool)
+        .await
+        .ok()?
+        .id,
+    )
+}
 
 #[derive(Serialize, Deserialize, Debug, Clone, TS)]
 #[serde(rename_all = "camelCase")]
 #[ts(export, export_to = "../web/src/types/")]
 pub struct SubFolderInsert {
     name: String,
-    parent_folder_id: i32
+    parent_folder_id: i32,
 }
-
-//TODO check user owns folder
 
 #[post("/folder/add")]
 pub async fn add_folder(
@@ -24,10 +46,16 @@ pub async fn add_folder(
     req: HttpRequest,
 ) -> impl Responder {
     let user_id: i32 = utils::get_user_id(&req).unwrap();
+    let owner = get_folder_owner(payload.parent_folder_id, data.db_pool.as_ref()).await;
+
+    if Some(user_id) != owner {
+        return HttpResponse::Unauthorized().body("User does not own that folder");
+    }
 
     sqlx::query!(
-        "INSERT INTO folder (name, parent_id) VALUES ($1, $2)", 
-        payload.name, payload.parent_folder_id
+        "INSERT INTO folder (name, parent_id) VALUES ($1, $2)",
+        payload.name,
+        payload.parent_folder_id
     )
     .execute(data.db_pool.as_ref())
     .await
@@ -35,3 +63,4 @@ pub async fn add_folder(
 
     HttpResponse::Ok().body("Folder insert successful")
 }
+
